* commit 'e4dcfc5d0dd7fd9aefc35126f2b6b41df03ff2fe': (15595 commits)
  Revert "drivers: hv: dxgkrnl: Make dxgkrnl to compile with new VMBUS interface"
  drivers: hv: dxgkrnl: Make dxgkrnl to compile with new VMBUS interface"
  drivers: hv: dxgkrnl: Do not print error messages when virtual GPU is not present
  drivers: hv: dxgkrnl: Use pin_user_pages instead of get_user_pages for DMA accessible memory
  drivers: hv: dxgkrnl: Implement D3DKMTEnumProcesses to match the Windows implementation
  drivers: hv: dxgkrnl: Add support for locking a shared allocation by not the owner
  drivers: hv: dxgkrnl: Retry sending a VM bus packet when there is no place in the ring buffer
  drivers: hv: dxgkrnl: Fixed the implementation of D3DKMTQueryClockCalibration
  drivers: hv: dxgkrnl: Fixed coding style issues
  drivers: hv: dxgkrnl: Implement known escapes
  drivers: hv: dxgkrnl: Implement D3DDKMTIsFeatureEnabled API
  drivers: hv: dxgkrnl: Implement the D3DKMTEnumProcesses API
  drivers: hv: dxgkrnl: Handle process ID in D3DKMTQueryStatistics
  drivers: hv: dxgkrnl: Added implementation for D3DKMTInvalidateCache
  drivers: hv: dxgkrnl: Added support for compute only adapters
  drivers: hv: dxgkrnl: Added missed NULL check for resource object
  drivers: hv: dxgkrnl: Close shared file objects in case of a failure
  drivers: hv: dxgkrnl: Fix synchronization locks
  drivers: hv: dxgkrnl: Improve tracing and return values from copy from user
  drivers: hv: dxgkrnl: Implement D3DKMTWaitSyncFile
  ...

+++ drivers/net/dsa/mv88e6xxx/serdes.c
@@ -338,8 +338,8 @@ static uint64_t mv88e6352_serdes_get_stat(struct mv88e6xxx_chip *chip,
 	return val;
 }
 
-int mv88e6352_serdes_get_stats(struct mv88e6xxx_chip *chip, int port,
-			       uint64_t *data)
+size_t mv88e6352_serdes_get_stats(struct mv88e6xxx_chip *chip, int port,
+				  uint64_t *data)
 {
 	struct mv88e6xxx_port *mv88e6xxx_port = &chip->ports[port];
 	struct mv88e6352_serdes_hw_stat *stat;
@@ -787,8 +787,8 @@ static uint64_t mv88e6390_serdes_get_stat(struct mv88e6xxx_chip *chip, int lane,
 	return reg[0] | ((u64)reg[1] << 16) | ((u64)reg[2] << 32);
 }
 
-int mv88e6390_serdes_get_stats(struct mv88e6xxx_chip *chip, int port,
-			       uint64_t *data)
+size_t mv88e6390_serdes_get_stats(struct mv88e6xxx_chip *chip, int port,
+				  uint64_t *data)
 {
 	struct mv88e6390_serdes_hw_stat *stat;
 	int lane;
